fix: ensure ctl-api post-deploy tasks run in correct order
We can't define or guarantee the order of execution for multiple actions
attached to the same hook. This PR combines all of the ctl-api
actions into a single pre-deploy and a single post-deploy.

The exception is the ch-init action. It's attached to ctl-api's
pre-deploy because it can take 20 minutes for clickhouse to be ready. We
don't want to add an additional 20 minutes to the byoc provision, so we
continue deploying while clickhouse spins up in the background. We don't
need clickhouse until ctl-api is deploying, so we attach it there.
Leaving that separate, since it does not need to run in order with any
of the other ctl-api actions.