feat: require custom domain (#153)
* fix: root domain input

I used to be able to leave the default value, and `nuon.run` domain
would automatically be provisioned. Now it's actually trying to use
`byoc.for.everyone.example.com`. I'm not sure what changed.

This commit removes the default value and nuon dns attribute, and
requires the user to provide a root domain.

* fix: spelling error

* feat: make required inputs required

* fix: clean up slack healthcheck steps